# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OMPAY Orange Money API",
        "description": "API pour l'application Orange Money - OMPAY",
        "contact": {
            "email": "support@orangemoney.sn"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement"
        },
        {
            "url": "https://api.ompaysn.com/api",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/paiement/scanner-qr": {
            "post": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Scanner un code QR",
                "description": "Scanne un code QR et récupère les informations du marchand",
                "operationId": "scannerQR",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "donneesQR"
                                ],
                                "properties": {
                                    "donneesQR": {
                                        "description": "Données du code QR au format OM_PAY_{idMarchand}_{montant}_{timestamp}_{signature}",
                                        "type": "string",
                                        "example": "OM_PAY_76104421-859e-4710-99d7-2800cefa5b0a_5000.00_1762861353_33f391d122cfedbee4cb4e3fc0a49166"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Données du QR décodées",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "idScan": {
                                                    "type": "string",
                                                    "example": "scn_aKm2WAosDn"
                                                },
                                                "marchand": {
                                                    "properties": {
                                                        "idMarchand": {
                                                            "type": "string",
                                                            "example": "MCH_VRMT7VQP"
                                                        },
                                                        "nom": {
                                                            "type": "string",
                                                            "example": "Boutique Orange"
                                                        },
                                                        "logo": {
                                                            "type": "string",
                                                            "example": "https://cdn.ompay.sn/logos/boutique_orange.png"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "montant": {
                                                    "type": "number",
                                                    "example": 22068
                                                },
                                                "devise": {
                                                    "type": "string",
                                                    "example": "XOF"
                                                },
                                                "dateExpiration": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-11-11T11:07:01+00:00"
                                                },
                                                "valide": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "QR Code scanné avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "QR Code invalide ou expiré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "PAYMENT_004"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "QR Code expiré"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/paiement/saisir-code": {
            "post": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Saisir un code de paiement",
                "description": "Saisir manuellement un code de paiement",
                "operationId": "saisirCode",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code"
                                ],
                                "properties": {
                                    "code": {
                                        "description": "Code de paiement alphanumérique",
                                        "type": "string",
                                        "example": "FRESH001"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Code validé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "idCode": {
                                                    "type": "string",
                                                    "example": "cod_NlqcI9RK6b"
                                                },
                                                "marchand": {
                                                    "properties": {
                                                        "idMarchand": {
                                                            "type": "string",
                                                            "example": "MCH_VRMT7VQP"
                                                        },
                                                        "nom": {
                                                            "type": "string",
                                                            "example": "Boutique Orange"
                                                        },
                                                        "logo": {
                                                            "type": "string",
                                                            "example": "https://cdn.ompay.sn/logos/boutique_orange.png"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "montant": {
                                                    "type": "number",
                                                    "example": 2500
                                                },
                                                "devise": {
                                                    "type": "string",
                                                    "example": "XOF"
                                                },
                                                "dateExpiration": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "nullable": true
                                                },
                                                "valide": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Code validé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Code invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "PAYMENT_006"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Code de paiement invalide"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}